services/sentinel-growth/requirements.txt
fastapi>=0.109.0
uvicorn[standard]
google-cloud-storage
google-cloud-iam
google-generativeai>=0.8.3
weasyprint>=60.0
docxtpl>=0.16.0
structlog
pydantic-settings
python-multipart
jinja2
requests
feedparser>=6.0.0
google-cloud-firestore
httpx>=0.27.0
pypdf
pandas
apscheduler>=3.10.0
firebase-admin>=6.4.0
google-cloud-pubsub>=2.23.0
resend>=2.0.0
